requirements.txt
mcp
google-genai
tree-sitter
tree-sitter-java
JPype1
python-dotenv